feat: move some props to runtime config
