+++ packages/core/src/core/diff-viewer/module.ts
@@ -2,7 +2,6 @@ import { BrowserModule } from '@opensumi/ide-core-browser';
 
 import { Injectable } from '@opensumi/di';
 import { DiffViewerContribution } from './internal/base';
-import { DiffViewerComponentContribution } from './internal/component.provider';
 import { DiffViewerThemeProvider } from './internal/theme.provider';
 
 @Injectable()
@@ -10,6 +9,5 @@ export class DiffViewerModule extends BrowserModule {
   providers = [
     DiffViewerContribution,
     DiffViewerThemeProvider,
-    DiffViewerComponentContribution,
   ];
 }
