+++ packages/core/src/core/internal/codeblitz.module.ts
@@ -1,11 +1,14 @@
 import { RuntimeConfig } from '@codeblitzjs/ide-sumi-core';
 import { Autowired, Injectable, Provider } from '@opensumi/di';
 import { BrowserModule, Domain, KeybindingContribution, KeybindingRegistry } from '@opensumi/ide-core-browser';
+import { ComponentContribution, ComponentRegistry } from '@opensumi/ide-core-browser';
+import { TabbarRightExtraContentId } from '@opensumi/ide-editor';
+import { ITheme, ThemeContributionProvider } from '@opensumi/ide-theme';
 import { CodeBlitzCommandContribution } from '../commands';
 import { ExtensionActivateContribution } from '../extension/extension.contribution';
 
-@Domain(KeybindingContribution)
-class CodeBlitzContribution implements KeybindingContribution {
+@Domain(KeybindingContribution, ThemeContributionProvider, ComponentContribution)
+class CodeBlitzContribution implements KeybindingContribution, ThemeContributionProvider, ComponentContribution {
   @Autowired(RuntimeConfig)
   private readonly runtimeConfig: RuntimeConfig;
 
@@ -20,6 +23,20 @@ class CodeBlitzContribution implements KeybindingContribution {
     if (this.runtimeConfig.registerKeybindings?.length) {
       keybindings.registerKeybindings(this.runtimeConfig.registerKeybindings);
     }
+  }
+
+  registerComponent(registry: ComponentRegistry) {
+    if (this.runtimeConfig.tabBarRightExtraContent) {
+      registry.register(TabbarRightExtraContentId, {
+        id: TabbarRightExtraContentId,
+        component: this.runtimeConfig.tabBarRightExtraContent.component,
+        initialProps: this.runtimeConfig.tabBarRightExtraContent.initialProps,
+      });
+    }
+  }
+
+  onWillApplyTheme(theme: ITheme): Record<string, string | undefined> {
+    return this.runtimeConfig?.onWillApplyTheme?.(theme) || {};
   }
 }
